backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
spacy
transformers
torch
python-multipart
PyPDF2
python-docx
